feat: user can write their own description for start operator node input field
