+++ packages/toolkit/src/lib/use-instill-form/components/start-operator-free-form-fields/AudiosField.tsx
@@ -8,6 +8,7 @@ export const AudiosField = ({
   form,
   path,
   title,
+  description,
   onEditField,
   onDeleteField,
 }: {
@@ -75,6 +76,7 @@ export const AudiosField = ({
                 />
               </Input.Root>
             </Form.Control>
+            <Form.Description text={description} />
             <Form.Message />
           </Form.Item>
         );
