+++ packages/toolkit/src/lib/use-instill-form/components/start-operator-free-form-fields/TextField.tsx
@@ -6,6 +6,7 @@ export const TextField = ({
   form,
   path,
   title,
+  description,
   onEditField,
   onDeleteField,
 }: {
@@ -40,6 +41,7 @@ export const TextField = ({
                 />
               </Input.Root>
             </Form.Control>
+            <Form.Description text={description} />
             <Form.Message />
           </Form.Item>
         );
