+++ packages/toolkit/src/lib/use-instill-form/components/start-operator-free-form-fields/TextareaField.tsx
@@ -6,6 +6,7 @@ export const TextareaField = ({
   form,
   path,
   title,
+  description,
   onEditField,
   onDeleteField,
 }: {
@@ -37,6 +38,7 @@ export const TextareaField = ({
                 className="!resize-y !text-[#1D2433] !text-opacity-80 !product-body-text-3-regular focus-visible:!ring-1"
               />
             </Form.Control>
+            <Form.Description text={description} />
             <Form.Message />
           </Form.Item>
         );
