+++ packages/toolkit/src/lib/use-instill-form/components/start-operator-free-form-fields/index.ts
@@ -1,8 +1,8 @@
 import { AudioField } from "./AudioField";
 import { AudiosField } from "./AudiosField";
 import { BooleanField } from "./BooleanField";
-import { ImageField } from "./NewImageField";
-import { ImagesField } from "./NewImagesField";
+import { ImageField } from "./ImageField";
+import { ImagesField } from "./ImagesField";
 import { LongTextField } from "./LongTextField";
 import { NumberField } from "./NumberField";
 import { NumbersField } from "./NumbersField";
